requirements.txt
gmusicapi>=7.0.0
blessed>=1.12.0
pygobject>=3.14.0
